Convenience function to apply descr to a list of datasets and print the results

Convenience function to apply descr to a list of datasets and print the results

lapply_descr(list, ...)

Arguments

list

a list of datasets (tibbles or data.frames)

...

arguments to be passed to the descr call

Value

something printable.

Examples

l <- list()
for (i in 1:2){
  l <- append(l, list(iris))
}
lapply_descr(l, group="Species")
#>    Variables      setosa       versicolor   virginica    Total       p     Test 
#>  1  Sepal.Length                                                                
#>  2    N            50           50           50           150         <0.…  F-t…
#>  3    mean         5            5.9          6.6          5.8                   
#>  4    sd           0.35         0.52         0.64         0.83                  
#>  5    median       5            5.9          6.5          5.8                   
#>  6    Q1 - Q3      4.8 -- 5.2   5.6 -- 6.3   6.2 -- 6.9   5.1 -- 6.…            
#>  7    min - max    4.3 -- 5.8   4.9 -- 7     4.9 -- 7.9   4.3 -- 7.…            
#>  8  Sepal.Width                                                                 
#>  9    N            50           50           50           150         <0.…  F-t…
#> 10    mean         3.4          2.8          3            3.1                   
#> 11    sd           0.38         0.31         0.32         0.44                  
#> 12    median       3.4          2.8          3            3                     
#> 13    Q1 - Q3      3.2 -- 3.7   2.5 -- 3     2.8 -- 3.2   2.8 -- 3.…            
#> 14    min - max    2.3 -- 4.4   2 -- 3.4     2.2 -- 3.8   2 -- 4.4              
#> 15  Petal.Length                                                                
#> 16    N            50           50           50           150         <0.…  F-t…
#> 17    mean         1.5          4.3          5.6          3.8                   
#> 18    sd           0.17         0.47         0.55         1.8                   
#> 19    median       1.5          4.3          5.5          4.3                   
#> 20    Q1 - Q3      1.4 -- 1.6   4 -- 4.6     5.1 -- 5.9   1.6 -- 5.…            
#> 21    min - max    1 -- 1.9     3 -- 5.1     4.5 -- 6.9   1 -- 6.9              
#> 22  Petal.Width                                                                 
#> 23    N            50           50           50           150         <0.…  F-t…
#> 24    mean         0.25         1.3          2            1.2                   
#> 25    sd           0.11         0.2          0.27         0.76                  
#> 26    median       0.2          1.3          2            1.3                   
#> 27    Q1 - Q3      0.2 -- 0.3   1.2 -- 1.5   1.8 -- 2.3   0.3 -- 1.…            
#> 28    min - max    0.1 -- 0.6   1 -- 1.8     1.4 -- 2.5   0.1 -- 2.…            
#>    Variables      setosa       versicolor   virginica    Total       p     Test 
#>  1  Sepal.Length                                                                
#>  2    N            50           50           50           150         <0.…  F-t…
#>  3    mean         5            5.9          6.6          5.8                   
#>  4    sd           0.35         0.52         0.64         0.83                  
#>  5    median       5            5.9          6.5          5.8                   
#>  6    Q1 - Q3      4.8 -- 5.2   5.6 -- 6.3   6.2 -- 6.9   5.1 -- 6.…            
#>  7    min - max    4.3 -- 5.8   4.9 -- 7     4.9 -- 7.9   4.3 -- 7.…            
#>  8  Sepal.Width                                                                 
#>  9    N            50           50           50           150         <0.…  F-t…
#> 10    mean         3.4          2.8          3            3.1                   
#> 11    sd           0.38         0.31         0.32         0.44                  
#> 12    median       3.4          2.8          3            3                     
#> 13    Q1 - Q3      3.2 -- 3.7   2.5 -- 3     2.8 -- 3.2   2.8 -- 3.…            
#> 14    min - max    2.3 -- 4.4   2 -- 3.4     2.2 -- 3.8   2 -- 4.4              
#> 15  Petal.Length                                                                
#> 16    N            50           50           50           150         <0.…  F-t…
#> 17    mean         1.5          4.3          5.6          3.8                   
#> 18    sd           0.17         0.47         0.55         1.8                   
#> 19    median       1.5          4.3          5.5          4.3                   
#> 20    Q1 - Q3      1.4 -- 1.6   4 -- 4.6     5.1 -- 5.9   1.6 -- 5.…            
#> 21    min - max    1 -- 1.9     3 -- 5.1     4.5 -- 6.9   1 -- 6.9              
#> 22  Petal.Width                                                                 
#> 23    N            50           50           50           150         <0.…  F-t…
#> 24    mean         0.25         1.3          2            1.2                   
#> 25    sd           0.11         0.2          0.27         0.76                  
#> 26    median       0.2          1.3          2            1.3                   
#> 27    Q1 - Q3      0.2 -- 0.3   1.2 -- 1.5   1.8 -- 2.3   0.3 -- 1.…            
#> 28    min - max    0.1 -- 0.6   1 -- 1.8     1.4 -- 2.5   0.1 -- 2.…