S3 override for print function for DescrList objects.

This function takes a DescrList object and converts it to eihter a DescrPrintCharacter or DescrPrintNumeric object, depending on the printFormat option. This object is then printed in an appropriate format.

# S3 method for DescrList
print(
  x,
  printFormat = options("DescrTabFormat")[[1]],
  silent = F,
  var_options = list(),
  group_labels = list(),
  format_p = NULL,
  format_summary_stats = list(),
  format_options = list(),
  ...
)

Arguments

x
A DescrList object returned from descr.

printFormat
Possible values: "console" (default), "tex", "html", "word", "numeric"

silent
I TRUE, suppresses output to stdout.

var_options
A named list of named lists. For each variable, you can have special options that apply only to that variable. These options are specified in this argument. See the details and examples for more explanation.

group_labels
named list of labels for the levels of the group variable in dat.

format_p
formatting function for p-values.

format_summary_stats
named list of formatting functions for summary statistics.

format_options
named list of formatting options.

...
further arguments to be passed along to print method

Value

A DescrPrint object which can be printed in various formats.

Details

There is no way to convert between DescrPrintCharacter and DescrPrintNumeric objects. The first type is for what you would usually want, the second type is mostly for debugging purposes. A DescrPrintCharacter object can be printed as html, tex code, as a flextable object or simply to the console.

Examples

descr(iris)

#>    Variables      Total        p        Test                              
#>  1  Sepal.Length                                                          
#>  2    N            150          <0.001   Students one-sample t-test       
#>  3    mean         5.8                                                    
#>  4    sd           0.83                                                   
#>  5    median       5.8                                                    
#>  6    Q1 - Q3      5.1 -- 6.4                                             
#>  7    min - max    4.3 -- 7.9                                             
#>  8  Sepal.Width                                                           
#>  9    N            150          <0.001   Students one-sample t-test       
#> 10    mean         3.1                                                    
#> 11    sd           0.44                                                   
#> 12    median       3                                                      
#> 13    Q1 - Q3      2.8 -- 3.3                                             
#> 14    min - max    2 -- 4.4                                               
#> 15  Petal.Length                                                          
#> 16    N            150          <0.001   Students one-sample t-test       
#> 17    mean         3.8                                                    
#> 18    sd           1.8                                                    
#> 19    median       4.3                                                    
#> 20    Q1 - Q3      1.6 -- 5.1                                             
#> 21    min - max    1 -- 6.9                                               
#> 22  Petal.Width                                                           
#> 23    N            150          <0.001   Students one-sample t-test       
#> 24    mean         1.2                                                    
#> 25    sd           0.76                                                   
#> 26    median       1.3                                                    
#> 27    Q1 - Q3      0.3 -- 1.8                                             
#> 28    min - max    0.1 -- 2.5                                             
#> 29  Species                                                               
#> 30    setosa       50 (33%)     >0.999   Chi-squared goodness-of-fit test 
#> 31    versicolor   50 (33%)                                               
#> 32    virginica    50 (33%)                                               

DescrPrint <- print(descr(iris))

#>    Variables      Total        p        Test                              
#>  1  Sepal.Length                                                          
#>  2    N            150          <0.001   Students one-sample t-test       
#>  3    mean         5.8                                                    
#>  4    sd           0.83                                                   
#>  5    median       5.8                                                    
#>  6    Q1 - Q3      5.1 -- 6.4                                             
#>  7    min - max    4.3 -- 7.9                                             
#>  8  Sepal.Width                                                           
#>  9    N            150          <0.001   Students one-sample t-test       
#> 10    mean         3.1                                                    
#> 11    sd           0.44                                                   
#> 12    median       3                                                      
#> 13    Q1 - Q3      2.8 -- 3.3                                             
#> 14    min - max    2 -- 4.4                                               
#> 15  Petal.Length                                                          
#> 16    N            150          <0.001   Students one-sample t-test       
#> 17    mean         3.8                                                    
#> 18    sd           1.8                                                    
#> 19    median       4.3                                                    
#> 20    Q1 - Q3      1.6 -- 5.1                                             
#> 21    min - max    1 -- 6.9                                               
#> 22  Petal.Width                                                           
#> 23    N            150          <0.001   Students one-sample t-test       
#> 24    mean         1.2                                                    
#> 25    sd           0.76                                                   
#> 26    median       1.3                                                    
#> 27    Q1 - Q3      0.3 -- 1.8                                             
#> 28    min - max    0.1 -- 2.5                                             
#> 29  Species                                                               
#> 30    setosa       50 (33%)     >0.999   Chi-squared goodness-of-fit test 
#> 31    versicolor   50 (33%)                                               
#> 32    virginica    50 (33%)                                               

DescrPrint$variables$Sepal.Length$Total$mean

#> [1] "5.8"

print(DescrPrint)

#> $variables
#> $variables$Sepal.Length
#> $Total
#> $Total$N
#> [1] "150"
#> 
#> $Total$mean
#> [1] "5.8"
#> 
#> $Total$sd
#> [1] "0.83"
#> 
#> $Total$median
#> [1] "5.8"
#> 
#> $Total$`Q1 - Q3`
#> [1] "5.1 -- 6.4"
#> 
#> $Total$`min - max`
#> [1] "4.3 -- 7.9"
#> 
#> 
#> $test_list
#> $test_list$p
#> [1] 3.331256e-129
#> 
#> $test_list$test_name
#> [1] "Students one-sample t-test"
#> 
#> 
#> $variable_name
#> [1] "Sepal.Length"
#> 
#> $variable_options
#> $variable_options$label
#> [1] "Sepal.Length"
#> 
#> 
#> $variable_lengths
#> $variable_lengths$Total
#> $variable_lengths$Total$N
#> [1] 150
#> 
#> $variable_lengths$Total$Nmiss
#> [1] 0
#> 
#> 
#> 
#> $label
#> [1] "Sepal.Length"
#> 
#> attr(,"class")
#> [1] "cont_summary" "list"        
#> 
#> $variables$Sepal.Width
#> $Total
#> $Total$N
#> [1] "150"
#> 
#> $Total$mean
#> [1] "3.1"
#> 
#> $Total$sd
#> [1] "0.44"
#> 
#> $Total$median
#> [1] "3"
#> 
#> $Total$`Q1 - Q3`
#> [1] "2.8 -- 3.3"
#> 
#> $Total$`min - max`
#> [1] "2 -- 4.4"
#> 
#> 
#> $test_list
#> $test_list$p
#> [1] 8.004458e-129
#> 
#> $test_list$test_name
#> [1] "Students one-sample t-test"
#> 
#> 
#> $variable_name
#> [1] "Sepal.Width"
#> 
#> $variable_options
#> $variable_options$label
#> [1] "Sepal.Width"
#> 
#> 
#> $variable_lengths
#> $variable_lengths$Total
#> $variable_lengths$Total$N
#> [1] 150
#> 
#> $variable_lengths$Total$Nmiss
#> [1] 0
#> 
#> 
#> 
#> $label
#> [1] "Sepal.Width"
#> 
#> attr(,"class")
#> [1] "cont_summary" "list"        
#> 
#> $variables$Petal.Length
#> $Total
#> $Total$N
#> [1] "150"
#> 
#> $Total$mean
#> [1] "3.8"
#> 
#> $Total$sd
#> [1] "1.8"
#> 
#> $Total$median
#> [1] "4.3"
#> 
#> $Total$`Q1 - Q3`
#> [1] "1.6 -- 5.1"
#> 
#> $Total$`min - max`
#> [1] "1 -- 6.9"
#> 
#> 
#> $test_list
#> $test_list$p
#> [1] 2.166017e-57
#> 
#> $test_list$test_name
#> [1] "Students one-sample t-test"
#> 
#> 
#> $variable_name
#> [1] "Petal.Length"
#> 
#> $variable_options
#> $variable_options$label
#> [1] "Petal.Length"
#> 
#> 
#> $variable_lengths
#> $variable_lengths$Total
#> $variable_lengths$Total$N
#> [1] 150
#> 
#> $variable_lengths$Total$Nmiss
#> [1] 0
#> 
#> 
#> 
#> $label
#> [1] "Petal.Length"
#> 
#> attr(,"class")
#> [1] "cont_summary" "list"        
#> 
#> $variables$Petal.Width
#> $Total
#> $Total$N
#> [1] "150"
#> 
#> $Total$mean
#> [1] "1.2"
#> 
#> $Total$sd
#> [1] "0.76"
#> 
#> $Total$median
#> [1] "1.3"
#> 
#> $Total$`Q1 - Q3`
#> [1] "0.3 -- 1.8"
#> 
#> $Total$`min - max`
#> [1] "0.1 -- 2.5"
#> 
#> 
#> $test_list
#> $test_list$p
#> [1] 2.659021e-42
#> 
#> $test_list$test_name
#> [1] "Students one-sample t-test"
#> 
#> 
#> $variable_name
#> [1] "Petal.Width"
#> 
#> $variable_options
#> $variable_options$label
#> [1] "Petal.Width"
#> 
#> 
#> $variable_lengths
#> $variable_lengths$Total
#> $variable_lengths$Total$N
#> [1] 150
#> 
#> $variable_lengths$Total$Nmiss
#> [1] 0
#> 
#> 
#> 
#> $label
#> [1] "Petal.Width"
#> 
#> attr(,"class")
#> [1] "cont_summary" "list"        
#> 
#> $variables$Species
#> $Total
#> $Total$setosa
#> [1] "50 (33%)"
#> 
#> $Total$versicolor
#> [1] "50 (33%)"
#> 
#> $Total$virginica
#> [1] "50 (33%)"
#> 
#> 
#> $test_list
#> $test_list$p
#> [1] 1
#> 
#> $test_list$test_name
#> [1] "Chi-squared goodness-of-fit test"
#> 
#> 
#> $variable_name
#> [1] "Species"
#> 
#> $variable_levels
#> [1] "setosa"     "versicolor" "virginica" 
#> 
#> $variable_options
#> $variable_options$label
#> [1] "Species"
#> 
#> 
#> $variable_lengths
#> $variable_lengths$Total
#> $variable_lengths$Total$N
#> [1] 150
#> 
#> $variable_lengths$Total$Nmiss
#> [1] 0
#> 
#> 
#> 
#> $label
#> [1] "Species"
#> 
#> attr(,"class")
#> [1] "cat_summary" "list"       
#> 
#> 
#> $lengths
#> $lengths$Sepal.Length
#> [1] 7
#> 
#> $lengths$Sepal.Width
#> [1] 7
#> 
#> $lengths$Petal.Length
#> [1] 7
#> 
#> $lengths$Petal.Width
#> [1] 7
#> 
#> $lengths$Species
#> [1] 4
#> 
#> 
#> $group_n
#> [1] 150
#> 
#> $group_names
#> [1] "Total"
#> 
#> $group_labels
#> [1] "Total"
#> 
#> $labels
#> $labels$Sepal.Length
#> [1] "Sepal.Length"
#> 
#> $labels$Sepal.Width
#> [1] "Sepal.Width"
#> 
#> $labels$Petal.Length
#> [1] "Petal.Length"
#> 
#> $labels$Petal.Width
#> [1] "Petal.Width"
#> 
#> $labels$Species
#> [1] "Species"
#> 
#> 
#> $tibble
#> # A tibble: 32 x 4
#>    Variables    Total        p        Test                        
#>    <chr>        <chr>        <chr>    <chr>                       
#>  1 Sepal.Length ""           ""       ""                          
#>  2 N            "150"        "<0.001" "Students one-sample t-test"
#>  3 mean         "5.8"        ""       ""                          
#>  4 sd           "0.83"       ""       ""                          
#>  5 median       "5.8"        ""       ""                          
#>  6 Q1 - Q3      "5.1 -- 6.4" ""       ""                          
#>  7 min - max    "4.3 -- 7.9" ""       ""                          
#>  8 Sepal.Width  ""           ""       ""                          
#>  9 N            "150"        "<0.001" "Students one-sample t-test"
#> 10 mean         "3.1"        ""       ""                          
#> # … with 22 more rows
#> 
#> attr(,"class")
#> [1] "DescrPrint"          "DescrPrintCharacter" "list"