S3 override for print function for DescrPrint objects

S3 override for print function for DescrPrint objects

Usage

# S3 method for DescrPrint
print(x, print_format = options("print_format")[[1]], silent = FALSE, ...)

Arguments

x

A DescrList object returned from descr.

print_format

Possible values: "console" (default), "tex", "html", "word", "numeric"

silent

I TRUE, suppresses output to stdout.

...

further arguments to be passed along to print method

Value

A DescrPrint object which can be printed in various formats.

Examples

descr(iris)
#>    Variable      Totl         p        Test  
#>  1  Sp.L                                     
#>  2    N           150          <0.001   tt1  
#>  3    mean        5.84                       
#>  4    sd          0.828                      
#>  5    median      5.8                        
#>  6    Q1 - Q3     5.1 -- 6.4                 
#>  7    min - max   4.3 -- 7.9                 
#>  8  Sp.W                                     
#>  9    N           150          <0.001   tt1  
#> 10    mean        3.06                       
#> 11    sd          0.436                      
#> 12    median      3                          
#> 13    Q1 - Q3     2.8 -- 3.3                 
#> 14    min - max   2 -- 4.4                   
#> 15  Pt.L                                     
#> 16    N           150          <0.001   tt1  
#> 17    mean        3.76                       
#> 18    sd          1.77                       
#> 19    median      4.35                       
#> 20    Q1 - Q3     1.6 -- 5.1                 
#> 21    min - max   1 -- 6.9                   
#> 22  Pt.W                                     
#> 23    N           150          <0.001   tt1  
#> 24    mean        1.2                        
#> 25    sd          0.762                      
#> 26    median      1.3                        
#> 27    Q1 - Q3     0.3 -- 1.8                 
#> 28    min - max   0.1 -- 2.5                 
#> 29  Spcs                                     
#> 30  sets          50 (33.3%)   >0.999   chi1 
#> 31  vrsc          50 (33.3%)                 
#> 32  vrgn          50 (33.3%)                 
DescrPrint <- print(descr(iris))
#>    Variable      Totl         p        Test  
#>  1  Sp.L                                     
#>  2    N           150          <0.001   tt1  
#>  3    mean        5.84                       
#>  4    sd          0.828                      
#>  5    median      5.8                        
#>  6    Q1 - Q3     5.1 -- 6.4                 
#>  7    min - max   4.3 -- 7.9                 
#>  8  Sp.W                                     
#>  9    N           150          <0.001   tt1  
#> 10    mean        3.06                       
#> 11    sd          0.436                      
#> 12    median      3                          
#> 13    Q1 - Q3     2.8 -- 3.3                 
#> 14    min - max   2 -- 4.4                   
#> 15  Pt.L                                     
#> 16    N           150          <0.001   tt1  
#> 17    mean        3.76                       
#> 18    sd          1.77                       
#> 19    median      4.35                       
#> 20    Q1 - Q3     1.6 -- 5.1                 
#> 21    min - max   1 -- 6.9                   
#> 22  Pt.W                                     
#> 23    N           150          <0.001   tt1  
#> 24    mean        1.2                        
#> 25    sd          0.762                      
#> 26    median      1.3                        
#> 27    Q1 - Q3     0.3 -- 1.8                 
#> 28    min - max   0.1 -- 2.5                 
#> 29  Spcs                                     
#> 30  sets          50 (33.3%)   >0.999   chi1 
#> 31  vrsc          50 (33.3%)                 
#> 32  vrgn          50 (33.3%)                 
DescrPrint$variables$results$Sepal.Length$Total$mean
#> NULL
print(DescrPrint)
#> $variables
#> $variables$Sepal.Length
#> $results
#> $results$Total
#> $results$Total$summary_stats
#> $results$Total$summary_stats$N
#> [1] "150"
#> 
#> $results$Total$summary_stats$mean
#> [1] "5.84"
#> 
#> $results$Total$summary_stats$sd
#> [1] "0.828"
#> 
#> $results$Total$summary_stats$median
#> [1] "5.8"
#> 
#> $results$Total$summary_stats$`Q1 - Q3`
#> [1] "5.1 -- 6.4"
#> 
#> $results$Total$summary_stats$`min - max`
#> [1] "4.3 -- 7.9"
#> 
#> 
#> $results$Total$categories
#> list()
#> 
#> 
#> 
#> $test_list
#> $test_list$p
#> [1] 3.331256e-129
#> 
#> $test_list$test_name
#> [1] "Student's one-sample t-test"
#> 
#> 
#> $variable_name
#> [1] "Sepal.Length"
#> 
#> $variable_options
#> $variable_options$label
#> [1] "Sepal.Length"
#> 
#> 
#> $variable_lengths
#> $variable_lengths$Total
#> $variable_lengths$Total$N
#> [1] 150
#> 
#> $variable_lengths$Total$Nmiss
#> [1] 0
#> 
#> 
#> 
#> $label
#> [1] "Sepal.Length"
#> 
#> attr(,"class")
#> [1] "summary_stats" "list"         
#> 
#> $variables$Sepal.Width
#> $results
#> $results$Total
#> $results$Total$summary_stats
#> $results$Total$summary_stats$N
#> [1] "150"
#> 
#> $results$Total$summary_stats$mean
#> [1] "3.06"
#> 
#> $results$Total$summary_stats$sd
#> [1] "0.436"
#> 
#> $results$Total$summary_stats$median
#> [1] "3"
#> 
#> $results$Total$summary_stats$`Q1 - Q3`
#> [1] "2.8 -- 3.3"
#> 
#> $results$Total$summary_stats$`min - max`
#> [1] "2 -- 4.4"
#> 
#> 
#> $results$Total$categories
#> list()
#> 
#> 
#> 
#> $test_list
#> $test_list$p
#> [1] 8.004458e-129
#> 
#> $test_list$test_name
#> [1] "Student's one-sample t-test"
#> 
#> 
#> $variable_name
#> [1] "Sepal.Width"
#> 
#> $variable_options
#> $variable_options$label
#> [1] "Sepal.Width"
#> 
#> 
#> $variable_lengths
#> $variable_lengths$Total
#> $variable_lengths$Total$N
#> [1] 150
#> 
#> $variable_lengths$Total$Nmiss
#> [1] 0
#> 
#> 
#> 
#> $label
#> [1] "Sepal.Width"
#> 
#> attr(,"class")
#> [1] "summary_stats" "list"         
#> 
#> $variables$Petal.Length
#> $results
#> $results$Total
#> $results$Total$summary_stats
#> $results$Total$summary_stats$N
#> [1] "150"
#> 
#> $results$Total$summary_stats$mean
#> [1] "3.76"
#> 
#> $results$Total$summary_stats$sd
#> [1] "1.77"
#> 
#> $results$Total$summary_stats$median
#> [1] "4.35"
#> 
#> $results$Total$summary_stats$`Q1 - Q3`
#> [1] "1.6 -- 5.1"
#> 
#> $results$Total$summary_stats$`min - max`
#> [1] "1 -- 6.9"
#> 
#> 
#> $results$Total$categories
#> list()
#> 
#> 
#> 
#> $test_list
#> $test_list$p
#> [1] 2.166017e-57
#> 
#> $test_list$test_name
#> [1] "Student's one-sample t-test"
#> 
#> 
#> $variable_name
#> [1] "Petal.Length"
#> 
#> $variable_options
#> $variable_options$label
#> [1] "Petal.Length"
#> 
#> 
#> $variable_lengths
#> $variable_lengths$Total
#> $variable_lengths$Total$N
#> [1] 150
#> 
#> $variable_lengths$Total$Nmiss
#> [1] 0
#> 
#> 
#> 
#> $label
#> [1] "Petal.Length"
#> 
#> attr(,"class")
#> [1] "summary_stats" "list"         
#> 
#> $variables$Petal.Width
#> $results
#> $results$Total
#> $results$Total$summary_stats
#> $results$Total$summary_stats$N
#> [1] "150"
#> 
#> $results$Total$summary_stats$mean
#> [1] "1.2"
#> 
#> $results$Total$summary_stats$sd
#> [1] "0.762"
#> 
#> $results$Total$summary_stats$median
#> [1] "1.3"
#> 
#> $results$Total$summary_stats$`Q1 - Q3`
#> [1] "0.3 -- 1.8"
#> 
#> $results$Total$summary_stats$`min - max`
#> [1] "0.1 -- 2.5"
#> 
#> 
#> $results$Total$categories
#> list()
#> 
#> 
#> 
#> $test_list
#> $test_list$p
#> [1] 2.659021e-42
#> 
#> $test_list$test_name
#> [1] "Student's one-sample t-test"
#> 
#> 
#> $variable_name
#> [1] "Petal.Width"
#> 
#> $variable_options
#> $variable_options$label
#> [1] "Petal.Width"
#> 
#> 
#> $variable_lengths
#> $variable_lengths$Total
#> $variable_lengths$Total$N
#> [1] 150
#> 
#> $variable_lengths$Total$Nmiss
#> [1] 0
#> 
#> 
#> 
#> $label
#> [1] "Petal.Width"
#> 
#> attr(,"class")
#> [1] "summary_stats" "list"         
#> 
#> $variables$Species
#> $results
#> $results$Total
#> $results$Total$summary_stats
#> list()
#> 
#> $results$Total$categories
#> $results$Total$categories$setosa
#> [1] "50 (33.3%)"
#> 
#> $results$Total$categories$versicolor
#> [1] "50 (33.3%)"
#> 
#> $results$Total$categories$virginica
#> [1] "50 (33.3%)"
#> 
#> 
#> 
#> 
#> $test_list
#> $test_list$p
#> [1] 1
#> 
#> $test_list$test_name
#> [1] "Chi-squared goodness-of-fit test"
#> 
#> 
#> $variable_name
#> [1] "Species"
#> 
#> $variable_levels
#> [1] "setosa"     "versicolor" "virginica" 
#> 
#> $variable_options
#> $variable_options$label
#> [1] "Species"
#> 
#> 
#> $variable_lengths
#> $variable_lengths$Total
#> $variable_lengths$Total$N
#> [1] 150
#> 
#> $variable_lengths$Total$Nmiss
#> [1] 0
#> 
#> 
#> 
#> $label
#> [1] "Species"
#> 
#> attr(,"class")
#> [1] "summary_stats" "list"         
#> 
#> 
#> $lengths
#> $lengths$Sepal.Length
#> [1] 7
#> 
#> $lengths$Sepal.Width
#> [1] 7
#> 
#> $lengths$Petal.Length
#> [1] 7
#> 
#> $lengths$Petal.Width
#> [1] 7
#> 
#> $lengths$Species
#> [1] 4
#> 
#> 
#> $group
#> $group$labels
#> list()
#> 
#> $group$lengths
#> Total 
#>   150 
#> 
#> 
#> $group_names
#> [1] "Total"
#> 
#> $format
#> $format$p
#> function (x) 
#> {
#>     pvalue(x, accuracy = accuracy, decimal.mark = decimal.mark, 
#>         prefix = prefix, add_p = add_p)
#> }
#> <bytecode: 0x55de2ad75240>
#> <environment: 0x55de33d09030>
#> 
#> $format$summary_stats
#> $format$summary_stats$N
#> function (x) 
#> {
#>     format(x, digits = 3, scientific = 4)
#> }
#> <environment: 0x55de34466520>
#> 
#> $format$summary_stats$mean
#> function (x) 
#> {
#>     format(x, digits = 3, scientific = 4)
#> }
#> <environment: 0x55de34466520>
#> 
#> $format$summary_stats$sd
#> function (x) 
#> {
#>     format(x, digits = 3, scientific = 4)
#> }
#> <environment: 0x55de34466520>
#> 
#> $format$summary_stats$median
#> function (x) 
#> {
#>     format(x, digits = 3, scientific = 4)
#> }
#> <environment: 0x55de34466520>
#> 
#> $format$summary_stats$Q1
#> function (x) 
#> {
#>     format(x, digits = 3, scientific = 4)
#> }
#> <environment: 0x55de34466520>
#> 
#> $format$summary_stats$Q3
#> function (x) 
#> {
#>     format(x, digits = 3, scientific = 4)
#> }
#> <environment: 0x55de34466520>
#> 
#> $format$summary_stats$min
#> function (x) 
#> {
#>     format(x, digits = 3, scientific = 4)
#> }
#> <environment: 0x55de34466520>
#> 
#> $format$summary_stats$max
#> function (x) 
#> {
#>     format(x, digits = 3, scientific = 4)
#> }
#> <environment: 0x55de34466520>
#> 
#> $format$summary_stats$CI
#> function (x) 
#> {
#>     format(x, digits = 3, scientific = 4)
#> }
#> <environment: 0x55de34466520>
#> 
#> 
#> $format$options
#> $format$options$print_Total
#> [1] TRUE
#> 
#> $format$options$print_p
#> [1] TRUE
#> 
#> $format$options$print_CI
#> [1] FALSE
#> 
#> $format$options$combine_mean_sd
#> [1] FALSE
#> 
#> $format$options$combine_median_Q1_Q3
#> [1] FALSE
#> 
#> $format$options$omit_factor_level
#> [1] "none"
#> 
#> $format$options$omit_Nmiss_if_0
#> [1] TRUE
#> 
#> $format$options$omit_missings_in_group
#> [1] TRUE
#> 
#> $format$options$percent_accuracy
#> [1] 0.1
#> 
#> $format$options$percent_suffix
#> [1] "%"
#> 
#> $format$options$row_percent
#> [1] FALSE
#> 
#> $format$options$Nmiss_row_percent
#> [1] FALSE
#> 
#> $format$options$absolute_relative_frequency_mode
#> [1] "both"          "only_absolute" "only_relative"
#> 
#> $format$options$omit_missings_in_categorical_var
#> [1] FALSE
#> 
#> $format$options$categorical_missing_percent_mode
#> [1] "no_missing_percent"           "missing_as_regular_category" 
#> [3] "missing_as_separate_category"
#> 
#> $format$options$caption
#> NULL
#> 
#> $format$options$replace_empty_string_with_NA
#> [1] TRUE
#> 
#> $format$options$categories_first_summary_stats_second
#> [1] FALSE
#> 
#> $format$options$max_first_col_width
#> [1] 7.5
#> 
#> 
#> $format$reshape_rows
#> $format$reshape_rows$`Q1 - Q3`
#> $format$reshape_rows$`Q1 - Q3`$args
#> [1] "Q1" "Q3"
#> 
#> $format$reshape_rows$`Q1 - Q3`$fun
#> function (Q1, Q3) 
#> {
#>     paste0(Q1, " -- ", Q3)
#> }
#> <environment: 0x55de34466520>
#> 
#> 
#> $format$reshape_rows$`min - max`
#> $format$reshape_rows$`min - max`$args
#> [1] "min" "max"
#> 
#> $format$reshape_rows$`min - max`$fun
#> function (min, max) 
#> {
#>     paste0(min, " -- ", max)
#> }
#> <environment: 0x55de34466520>
#> 
#> 
#> 
#> $format$test_abbreviations
#> $format$test_abbreviations$`Cochran's Q test`
#> [1] "CocQ"
#> 
#> $format$test_abbreviations$`McNemar's test`
#> [1] "McN"
#> 
#> $format$test_abbreviations$`Chi-squared goodness-of-fit test`
#> [1] "chi1"
#> 
#> $format$test_abbreviations$`Pearson's chi-squared test`
#> [1] "chi2"
#> 
#> $format$test_abbreviations$`Exact McNemar's test`
#> [1] "eMcN"
#> 
#> $format$test_abbreviations$`Boschloo's test`
#> [1] "Bolo"
#> 
#> $format$test_abbreviations$`Exact binomial test`
#> [1] "Bin"
#> 
#> $format$test_abbreviations$`Fisher's exact test`
#> [1] "Fish"
#> 
#> $format$test_abbreviations$`Friedman test`
#> [1] "Frie"
#> 
#> $format$test_abbreviations$`Wilcoxon two-sample signed-rank test`
#> [1] "Wil2"
#> 
#> $format$test_abbreviations$`Wilcoxon's one-sample signed-rank test`
#> [1] "Wil1"
#> 
#> $format$test_abbreviations$`Mann-Whitney's U test`
#> [1] "MWU"
#> 
#> $format$test_abbreviations$`Kruskal-Wallis's one-way ANOVA`
#> [1] "KW"
#> 
#> $format$test_abbreviations$`Student's paired t-test`
#> [1] "tpar"
#> 
#> $format$test_abbreviations$`Mixed model ANOVA`
#> [1] "MiAn"
#> 
#> $format$test_abbreviations$`Student's one-sample t-test`
#> [1] "tt1"
#> 
#> $format$test_abbreviations$`Student's two-sample t-test`
#> [1] "stt2"
#> 
#> $format$test_abbreviations$`Welch's two-sample t-test`
#> [1] "tt2"
#> 
#> $format$test_abbreviations$`F-test (ANOVA)`
#> [1] "F"
#> 
#> $format$test_abbreviations$`Cochran-Armitage's test`
#> [1] "CocA"
#> 
#> $format$test_abbreviations$`Jonckheere-Terpstra's test`
#> [1] "JT"
#> 
#> $format$test_abbreviations$`CI for difference in proportions derived from a normal ("Wald") approximation`
#> [1] "PWa"
#> 
#> $format$test_abbreviations$`CI for difference in proportions derived from an unconditional exact test`
#> [1] "PUnc"
#> 
#> $format$test_abbreviations$`CI for difference in proportions derived from an exact McNemar's test`
#> [1] "PMcN"
#> 
#> $format$test_abbreviations$`CI for difference in means derived from the t-distribution`
#> [1] "t"
#> 
#> $format$test_abbreviations$`CI for the Hodges-Lehmann estimator`
#> [1] "HL"
#> 
#> $format$test_abbreviations$`CI for odds ratio derived from Fisher's exact test`
#> [1] "Odds"
#> 
#> $format$test_abbreviations$`No test`
#> [1] "NA"
#> 
#> 
#> 
#> $group_labels
#> [1] "Total"
#> 
#> $labels
#> $labels$Sepal.Length
#> [1] "Sepal.Length"
#> 
#> $labels$Sepal.Width
#> [1] "Sepal.Width"
#> 
#> $labels$Petal.Length
#> [1] "Petal.Length"
#> 
#> $labels$Petal.Width
#> [1] "Petal.Width"
#> 
#> $labels$Species
#> [1] "Species"
#> 
#> 
#> $tibble
#> # A tibble: 32 × 4
#>    Variable     Total        p        Test                         
#>    <chr>        <chr>        <chr>    <chr>                        
#>  1 Sepal.Length ""           ""       ""                           
#>  2 N            "150"        "<0.001" "Student's one-sample t-test"
#>  3 mean         "5.84"       ""       ""                           
#>  4 sd           "0.828"      ""       ""                           
#>  5 median       "5.8"        ""       ""                           
#>  6 Q1 - Q3      "5.1 -- 6.4" ""       ""                           
#>  7 min - max    "4.3 -- 7.9" ""       ""                           
#>  8 Sepal.Width  ""           ""       ""                           
#>  9 N            "150"        "<0.001" "Student's one-sample t-test"
#> 10 mean         "3.06"       ""       ""                           
#> # ℹ 22 more rows
#> 
#> attr(,"class")
#> [1] "DescrPrint"          "DescrPrintCharacter" "list"