Convenience function to apply descr to a list of datasets and print the results

Convenience function to apply descr to a list of datasets and print the results

Usage

lapply_descr(list, ...)

Arguments

list

a list of datasets (tibbles or data.frames)

...

arguments to be passed to the descr call

Value

something printable.

Examples

l <- list()
for (i in 1:2){
  l <- append(l, list(iris))
}
lapply_descr(l, group="Species")
#>    Variable      sets         vrsc         vrgn         Totl         p     Test 
#>  1  Sp.L                                                                        
#>  2    N           50           50           50           150          <0.…  F   
#>  3    mean        5.01         5.94         6.59         5.84                   
#>  4    sd          0.352        0.516        0.636        0.828                  
#>  5    median      5            5.9          6.5          5.8                    
#>  6    Q1 - Q3     4.8 -- 5.2   5.6 -- 6.3   6.2 -- 6.9   5.1 -- 6.4             
#>  7    min - max   4.3 -- 5.8   4.9 -- 7     4.9 -- 7.9   4.3 -- 7.9             
#>  8  Sp.W                                                                        
#>  9    N           50           50           50           150          <0.…  F   
#> 10    mean        3.43         2.77         2.97         3.06                   
#> 11    sd          0.379        0.314        0.322        0.436                  
#> 12    median      3.4          2.8          3            3                      
#> 13    Q1 - Q3     3.2 -- 3.7   2.5 -- 3     2.8 -- 3.2   2.8 -- 3.3             
#> 14    min - max   2.3 -- 4.4   2 -- 3.4     2.2 -- 3.8   2 -- 4.4               
#> 15  Pt.L                                                                        
#> 16    N           50           50           50           150          <0.…  F   
#> 17    mean        1.46         4.26         5.55         3.76                   
#> 18    sd          0.174        0.47         0.552        1.77                   
#> 19    median      1.5          4.35         5.55         4.35                   
#> 20    Q1 - Q3     1.4 -- 1.6   4 -- 4.6     5.1 -- 5.9   1.6 -- 5.1             
#> 21    min - max   1 -- 1.9     3 -- 5.1     4.5 -- 6.9   1 -- 6.9               
#> 22  Pt.W                                                                        
#> 23    N           50           50           50           150          <0.…  F   
#> 24    mean        0.246        1.33         2.03         1.2                    
#> 25    sd          0.105        0.198        0.275        0.762                  
#> 26    median      0.2          1.3          2            1.3                    
#> 27    Q1 - Q3     0.2 -- 0.3   1.2 -- 1.5   1.8 -- 2.3   0.3 -- 1.8             
#> 28    min - max   0.1 -- 0.6   1 -- 1.8     1.4 -- 2.5   0.1 -- 2.5             
#>    Variable      sets         vrsc         vrgn         Totl         p     Test 
#>  1  Sp.L                                                                        
#>  2    N           50           50           50           150          <0.…  F   
#>  3    mean        5.01         5.94         6.59         5.84                   
#>  4    sd          0.352        0.516        0.636        0.828                  
#>  5    median      5            5.9          6.5          5.8                    
#>  6    Q1 - Q3     4.8 -- 5.2   5.6 -- 6.3   6.2 -- 6.9   5.1 -- 6.4             
#>  7    min - max   4.3 -- 5.8   4.9 -- 7     4.9 -- 7.9   4.3 -- 7.9             
#>  8  Sp.W                                                                        
#>  9    N           50           50           50           150          <0.…  F   
#> 10    mean        3.43         2.77         2.97         3.06                   
#> 11    sd          0.379        0.314        0.322        0.436                  
#> 12    median      3.4          2.8          3            3                      
#> 13    Q1 - Q3     3.2 -- 3.7   2.5 -- 3     2.8 -- 3.2   2.8 -- 3.3             
#> 14    min - max   2.3 -- 4.4   2 -- 3.4     2.2 -- 3.8   2 -- 4.4               
#> 15  Pt.L                                                                        
#> 16    N           50           50           50           150          <0.…  F   
#> 17    mean        1.46         4.26         5.55         3.76                   
#> 18    sd          0.174        0.47         0.552        1.77                   
#> 19    median      1.5          4.35         5.55         4.35                   
#> 20    Q1 - Q3     1.4 -- 1.6   4 -- 4.6     5.1 -- 5.9   1.6 -- 5.1             
#> 21    min - max   1 -- 1.9     3 -- 5.1     4.5 -- 6.9   1 -- 6.9               
#> 22  Pt.W                                                                        
#> 23    N           50           50           50           150          <0.…  F   
#> 24    mean        0.246        1.33         2.03         1.2                    
#> 25    sd          0.105        0.198        0.275        0.762                  
#> 26    median      0.2          1.3          2            1.3                    
#> 27    Q1 - Q3     0.2 -- 0.3   1.2 -- 1.5   1.8 -- 2.3   0.3 -- 1.8             
#> 28    min - max   0.1 -- 0.6   1 -- 1.8     1.4 -- 2.5   0.1 -- 2.5